% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_obj.R
\name{barplot_obj}
\alias{barplot_obj}
\title{Aggreagting columns into "others" till given color.}
\usage{
barplot_obj(
  x,
  meta_cols = NULL,
  color = palettes(35, "rainbow"),
  othersCol = "grey30",
  specificName = NULL,
  specificColor = "grey90",
  sortFun = sum,
  na.rm = TRUE
)
}
\arguments{
\item{mat}{x is a data.frame of community matrix.}
}
\description{
Aggreagting columns into "others" till given color.
}
