% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.dir.R
\name{make.dir}
\alias{make.dir}
\title{Make directory to save}
\usage{
make.dir(file.name)
}
\arguments{
\item{file.name}{file.name is directory name.}
}
\description{
To tidy up your R directory.
}
\examples{
make.dir('output')

}
