% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeColPalette.R
\name{makeColPalette}
\alias{makeColPalette}
\title{Create a color correspondence table}
\usage{
makeColPalette(
  data,
  color = NULL,
  othersCol = "grey30",
  specificName = NULL,
  specificColor = "grey90",
  sortFun = sum,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A matrix or data frame containing samples × abundances.}

\item{color}{A vector containing color information.}

\item{specificName}{A species name for which a specific color should be assigned (e.g., "Unidentified").}

\item{specificColor}{Color assigned to the species specified in \code{specificName}.}

\item{others}{Color assigned to non-dominant categories.}
}
\description{
Creates a color correspondence table for ggplot2.
}
\examples{
data(test)
col <- palettes()
colpal <- makeColPalette(data=test[,-1],
                            color=col,
                            others='grey30',
                            specificName = 'unidentified',
                            specificColor = 'grey90',
                            sort=sum)

}
