% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palettes}
\alias{palettes}
\title{Generate a color palette}
\usage{
palettes(x = 70, pal = "RColorBrewer")
}
\arguments{
\item{x}{A single integer specifying the number of colors to generate. Default is 70.}

\item{pal}{Palette type to use. Choose from \code{"RColorBrewer"} or \code{"rainbow"}. Default is \code{"RColorBrewer"}.}
}
\description{
Generates up to 70 colors using palettes from RColorBrewer.
}
\examples{
palettes(10)
palettes(20)
palettes(20, pal='rainbow')

}
\references{
https://stackoverflow.com/questions/15282580/how-to-generate-a-number-of-most-distinctive-colors-in-r
}
