% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.mat.R
\name{Taxa.mat}
\alias{Taxa.mat}
\title{Aggregate an abundance matrix to another taxonomic level}
\usage{
Taxa.mat(
  data,
  taxa,
  taxaLabel,
  func = function(x) {
     sum(x)
 }
)
}
\arguments{
\item{data}{An abundance table in which columns represent ASVs/OTUs.}

\item{taxa}{A taxonomy table containing taxonomic information for each ASV/OTU.}

\item{taxaLabel}{A taxonomic level to which the abundance table should be aggregated.}

\item{func}{A function used for aggregation. The default is \code{sum()}.}
}
\description{
This function aggregates abundances to a specified taxonomic level, similar to \code{aggregate()}.
}
\examples{
# Aggregate an OTU table to the Family level
library(microbiome)
data(atlas1006)

data <- as.data.frame(t(otu_table(atlas1006)))
taxa <- tax_table(atlas1006)
taxaLabel <- "Family"
a <- Taxa.mat(data, taxa, taxaLabel)

}
